typo(cluster/manager): naming
